+++ python-impl/tests.py
@@ -1,11 +1,13 @@
 # flake8: noqa: E501
 
+import time
 from ec import (generator_Fq, generator_Fq2, default_ec, default_ec_twist,
                 y_for_x, hash_to_point_Fq, hash_to_point_Fq2,
                 twist, untwist, rand_scalar)
 from fields import Fq2, Fq6, Fq12, Fq
-from bls import BLSPrivateKey, BLS, AggregationInfo
-import time
+from bls import BLS
+from keys import BLSPrivateKey, BLSPublicKey, ExtendedPrivateKey
+from aggregation_info import AggregationInfo
 
 def test_fields():
     a = Fq(17, 30)
@@ -141,6 +143,28 @@ def test_vectors2():
     assert(BLS.verify(sig_final))
 
 
+def test_vectors3():
+    seed = bytes([1, 50, 6, 244, 24, 199, 1, 25])
+    esk =  ExtendedPrivateKey.from_seed(seed)
+    assert(esk.private_key.get_public_key().get_fingerprint() == 0xa4700b27)
+    assert(esk.chain_code.hex() == "d8b12555b4cc5578951e4a7c80031e22019cc0dce168b3ed88115311b8feb1e3")
+    esk77 = esk.private_child(77 + 2**31)
+    assert(esk77.chain_code.hex() == "f2c8e4269bb3e54f8179a5c6976d92ca14c3260dd729981e9d15f53049fd698b")
+    assert(esk77.private_key.get_public_key().get_fingerprint() == 0xa8063dcf)
+
+    assert(esk.private_child(3)
+              .private_child(17)
+              .private_key
+              .get_public_key()
+              .get_fingerprint() == 0xff26a31f)
+
+    assert(esk.get_extended_public_key()
+              .public_child(3)
+              .public_child(17)
+              .get_public_key()
+              .get_fingerprint() == 0xff26a31f)
+
+
 def test_bls():
     sk = BLSPrivateKey(rand_scalar())
     sk2 = BLSPrivateKey(rand_scalar())
@@ -159,6 +183,7 @@ test_ec()
 test_bls()
 test_vectors()
 test_vectors2()
+test_vectors3()
 
 
 """
