+++ python-impl/util.py
@@ -26,6 +26,23 @@ def hmac256(m, k):
     return hash256(kopad + hash256(kipad + m))
 
 
+def hash_pks(num_outputs, public_keys):
+    """
+    Construction from https://eprint.iacr.org/2018/483.pdf
+    Two hashes are performed for speed.
+    """
+    input_bytes = b''.join([pk.serialize() for pk in public_keys])
+    pk_hash = hash256(input_bytes)
+    order = public_keys[0].value.ec.n
+
+    computed_Ts = []
+    for i in range(num_outputs):
+        t = int.from_bytes(hash256(i.to_bytes(4, "big") + pk_hash), "big")
+        computed_Ts.append(t % order)
+
+    return computed_Ts
+
+
 """
 Copyright 2018 Chia Network Inc
